refactor: change README.md
- Change name of posts:  src/features/post to src/features/posts
