+++ README.md
@@ -1,40 +1,84 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+
+# Photo-Fiesta
+
+Photo-Fiesta is an Instagram clone that offers a comprehensive social media experience. Built with a microservices architecture using the Feature-Sliced Design (FSD) approach, this project showcases modern web development practices and technologies. Visit the live version at [https://photo-fiesta.online/](https://photo-fiesta.online/).
+
+## Features
+
+- **Authentication**: Secure login with Google and GitHub OAuth2, JWT and Refresh Tokens
+- **User Profiles**: Customizable profiles with activity feeds
+- **Posts**: Create, edit, delete, and interact with posts
+- **Media Sharing**: Upload and share photos
+- **Interactions**: Like and comment on posts
+- **Stories**: Share temporary content
+- **Messaging**: Real-time chat using WebSockets
+- **Statistics**: Track account performance
+- **Payment Integration**: Secure payment system for premium features
+
+## Technologies
+
+- TypeScript, Next.js, Redux Toolkit (RTK Query)
+- React Hook Form, Custom UI library [@photo-fiesta/ui-lib](https://www.npmjs.com/package/@photo-fiesta/ui-lib)
+- Docker, Jenkins, Husky, Vite
+- SASS, ESLint, Prettier, Stylelint
 
 ## Getting Started
 
-First, run the development server:
+To run Photo-Fiesta locally:
+
+1. Clone the repository
+2. Install dependencies: `pnpm install`
+3. Set up environment variables
+4. Run the development server: `pnpm run dev`
+
+Visit `http://localhost:3000` to view the application.
+
+## Development
+
+### Requirements
+
+Node.js v14+ and pnpm are required to install and run the project.
+
+### Installing Dependencies
+
+To install dependencies, run:
 
 ```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
+pnpm install
 ```
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+### Running Development Server
 
-You can start editing the page by modifying `pages/[userId].tsx`. The page auto-updates as you edit the file.
+To start the development server, execute:
 
-[API routes](https://nextjs.org/docs/api-routes/introduction) can be accessed on [http://localhost:3000/api/hello](http://localhost:3000/api/hello). This endpoint can be edited in `pages/api/hello.ts`.
+```bash
+pnpm run dev
+```
 
-The `pages/api` directory is mapped to `/api/*`. Files in this directory are treated as [API routes](https://nextjs.org/docs/api-routes/introduction) instead of React pages.
+### Creating a Production Build
 
-This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
+To create a production build, use:
 
-## Learn More
+```bash
+pnpm run build
+```
 
-To learn more about Next.js, take a look at the following resources:
+### Running Tests
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+To run the test suite, use:
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
+```bash
+pnpm run test
+```
 
-## Deploy on Vercel
+## Deployment
 
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
+The application is deployed at [https://photo-fiesta.online/](https://photo-fiesta.online/). For detailed deployment instructions, please refer to our deployment documentation.
 
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+## Contributing
+
+We welcome contributions! Please read our contributing guidelines before submitting pull requests.
+
+## License
+
+This project is licensed under the MIT License.
