+++ src/features/index.ts
@@ -1,3 +1,3 @@
 export * from './auth'
-export * from './post'
+export * from './posts'
 export * from './profile'
