+++ src/features/posts/ui/index.ts
