+++ src/features/posts/ui/post/post.tsx
