+++ src/features/posts/ui/postList/index.ts
